Demos update CmakeLists.txt

+++ demos/common/cpp/models/src/detection_model_faceboxes.cpp
@@ -146,8 +146,8 @@ void ModelFaceBoxes::priorBoxes(const std::vector<std::pair<size_t, size_t>>& fe
 
     for (size_t k = 0; k < featureMaps.size(); ++k) {
         std::vector<float> a;
-        for (int i = 0; i < featureMaps[k].first; ++i) {
-            for (int j = 0; j < featureMaps[k].second; ++j) {
+        for (size_t i = 0; i < featureMaps[k].first; ++i) {
+            for (size_t j = 0; j < featureMaps[k].second; ++j) {
                 if (k == 0) {
                     calculateAnchorsZeroLevel(anchors, j, i,  minSizes[k], steps[k]);;
                 }
